Refactor: separating main.py into locale_level_analysis.py (for county/state/zip level analysis of Zillow data) and property_level_analysis.py (for individual property level analysis from Rentcast/Rentometer data).
